ai_fitness_backend/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-fitness-backend"
version = "0.1.0"
description = "AI Fitness Backend"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
dependencies = [
    "fastapi",
    "uvicorn",
    "python-multipart",
    "numpy",
    "pandas",
    "requests",
    "python-dotenv",
    "sentence-transformers",
    "scikit-learn",
    "sqlalchemy",
    "psycopg2-binary",
    "black>=24.8.0",
    "scrapy>=2.11.2",
    "aiohttp>=3.10.11",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
